GPS decoder

+++ Core/Inc/GPSdecode.h
@@ -1,10 +1,10 @@
 
 
 typedef struct {
-    unsigned int   latitude_bd;
-    unsigned int   nshemi_bd;
-    unsigned int   longitude_bd;
-    unsigned int   ewhemi_bd;
+	uint32_t   latitude_bd;
+	uint8_t    nshemi_bd;
+	uint32_t   longitude_bd;
+	uint8_t    ewhemi_bd;
 } GPSmessage;
 
 typedef struct {
@@ -12,10 +12,10 @@ typedef struct {
 	float      Latitude;
 } GPS_Data_TypeDef;
 
-int NMEA_Comma_Pos (int *buf,uint8_t cx);
+uint8_t  NMEA_Comma_Pos (uint8_t *buf, uint8_t cx);
 
-int NMEA_Pow (int m, int n);
+uint32_t NMEA_Pow (uint8_t m, uint8_t n);
 
-int NMEA_Str2num (int *buf, int *dx);
+int NMEA_Str2num (uint8_t *buf, uint8_t *dx);
 
-void NMEA_BDS_GPRMC_Analysis (GPSmessage *gpsmsg, int *buf);
+void NMEA_BDS_GPRMC_Analysis (GPSmessage *gpsmsg, uint8_t *buf);
